# missahoje-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/usuarios": {
      "post": {
        "operationId": "UsuariosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cria um novo usuário",
        "tags": [
          "Usuários"
        ]
      },
      "get": {
        "operationId": "UsuariosController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UsuarioDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista todos os usuários paginados",
        "tags": [
          "Usuários"
        ]
      }
    },
    "/usuarios/{id}": {
      "get": {
        "operationId": "UsuariosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Busca um usuário pelo ID",
        "tags": [
          "Usuários"
        ]
      },
      "put": {
        "operationId": "UsuariosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualiza um usuário",
        "tags": [
          "Usuários"
        ]
      },
      "delete": {
        "operationId": "UsuariosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove um usuário",
        "tags": [
          "Usuários"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Missa Hoje",
    "description": "Documentação da API do projeto Missa Hoje",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "itemCount": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          }
        },
        "required": [
          "itemCount",
          "totalItems",
          "itemsPerPage",
          "totalPages",
          "currentPage"
        ]
      },
...
      "CreateUsuarioDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "senha": {
            "type": "string"
          },
          "cidade_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "nome",
          "email",
          "senha"
        ]
      },
      "UsuarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "cidade_id": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "role",
          "cidade_id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateUsuarioDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}